apps/tasks/models.py
from django.db import models
from django.utils.translation import ugettext_lazy as _
from model_utils.models import TimeStampedModel

__all__ = (
    'Tasks',
)


class Tasks(TimeStampedModel):
    """Tasks"""
    name = models.CharField(verbose_name=_('Name'), max_length=255)

    class Meta:
        verbose_name = _('Task')
        verbose_name_plural = _('Tasks')

    def __str__(self):
        return f'{self.name}'

