requirements.txt
gunicorn==20.0.0
django==2.2.7
psycopg2==2.8.4
pytz==2019.3
dj-database-url==0.5.0
django_heroku==0.3.1
djangorestframework==3.10.3
django-filters==0.2.1
django-model-utils==3.2.0
django-rest-auth==0.9.5
django-allauth==0.40.0